backend/python/mlx-video/requirements.txt
# Core dependencies for MLX Video Backend
grpcio>=1.60.0
grpcio-tools>=1.60.0
requests>=2.31.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.66.0
# Deep learning and video generation
torch>=2.1.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
# Video processing
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
